feat: Adding technical and soft skill procedures

+++ Back/Banco-de-Talentos/src/main/java/com/fractal/bancodetalentos/model/response/CurrenciesResp.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CurrenciesResp {
-    private BigDecimal id;
+    private Integer id;
     private String name;
     private String code;
 }
